+++ src/rs/lib.rs
@@ -20,6 +20,9 @@ pub use segment::Segment;
 pub use symbol::Symbol;
 pub use symbol_comparison_info::SymbolComparisonInfo;
 
+#[macro_use]
+extern crate lazy_static;
+
 #[pyo3::prelude::pymodule]
 fn mapfile_parser(
     _py: pyo3::prelude::Python<'_>,